refactor: simplify element finding logic
since there is no reliable way to detect problem changes (even using navigation API, the problem still loads after the navigation happens, so the observer will still retain the previous problem), and the element finding is relatively fast, this commit moves the element finding from on page load to on every click of the button.

Also refactor element finding to use promise, so we can `await` the element to be found before doing anything else.